scripts/requirements.txt
PyGithub
openai
requests
